refactor(chat-markdown): extract stream-syntax and tail-cursor into utils/
